fix bundle index issues, caused through restructuring some component and stories folders
