+++ package.json
@@ -7,7 +7,7 @@
     "build": "cd packages/design-system-component-library-yeahyeahyeah && yarn build",
     "build_local": "cd packages/design-system-component-library-yeahyeahyeah && yarn build_local",
     "build-app": "yarn build && cd packages/app && yarn build",
-    "build-icons": "npx @svgr/cli --typescript --icon 16px --out-dir ./packages/design-system-component-library-yeahyeahyeah/components/icons/components -- ./packages/design-system-component-library-yeahyeahyeah/components/icons/svg",
+    "build-icons": "npx @svgr/cli --typescript --icon 16px --out-dir ./packages/design-system-component-library-yeahyeahyeah/components/icon/components -- ./packages/design-system-component-library-yeahyeahyeah/components/icon/svg",
     "watch": "cd packages/design-system-component-library-yeahyeahyeah && yarn watch",
     "storybook": "cd packages/design-system-component-library-yeahyeahyeah && yarn storybook",
     "build-storybook": "cd packages/design-system-component-library-yeahyeahyeah && yarn build-storybook",
