+++ packages/design-system-component-library-yeahyeahyeah/.ctiignore
@@ -1,5 +1,5 @@
 {
-  "components/icons/components/**": "*",
+  "components/icon/components/**": "*",
   "node_modules/**": "*",
   "stories/**": "*",
   "docs/**": "*"
