+++ packages/design-system-component-library-yeahyeahyeah/components/image/index.ts
@@ -1,0 +1,1 @@
+export * from './ImageContainer';
