--- packages/design-system-component-library-yeahyeahyeah/components/index.ts
@@ -1,13 +1,0 @@
-export * from './user/Avatar';
-export * from './branding';
-export * from './button/index';
-export * from './layout/Container';
-export * from './form/index';
-export * from './link/IconLink';
-export * from './icon/index';
-export * from './image/ImageContainer';
-export * from './modal';
-export * from './mumble';
-export * from './navigation/Navigation';
-export * from './typography';
-export * from './user/User';
