+++ packages/design-system-component-library-yeahyeahyeah/components/link/index.ts
@@ -1,0 +1,1 @@
+export * from './IconLink';
