+++ packages/design-system-component-library-yeahyeahyeah/components/navigation/index.ts
@@ -1,0 +1,1 @@
+export * from './Navigation';
