+++ packages/design-system-component-library-yeahyeahyeah/components/user/index.ts
@@ -1,0 +1,2 @@
+export * from './Avatar';
+export * from './User';
