+++ packages/design-system-component-library-yeahyeahyeah/index.ts
@@ -1,3 +1,14 @@
-export * from './components';
+export * from './components/branding';
+export * from './components/button';
+export * from './components/form';
+export * from './components/icon';
+export * from './components/image';
+export * from './components/layout';
+export * from './components/link';
+export * from './components/modal';
+export * from './components/mumble';
+export * from './components/navigation';
+export * from './components/typography';
+export * from './components/user';
 export * from './styles';
 export * from './types';
